requirements.txt
fastapi
uvicorn
pydantic
alembic
sqlalchemy
python-jose[cryptography]
python-dotenv
pydantic-settings
passlib[bcrypt]
python-multipart